+++ CHANGELOG.md
@@ -645,3 +645,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Move rewards scrolling code from Hubris to BaseMod (kiooeht/erasels)
 
 #### dev ####
+* Make Scrawl work with max hand size changes (kiooeht)